Read in an integer, not a short. Fixes #76
Java does not have unsigned types. As such, too big of short values will
roll over. Account for this with an integer datatype.

+++ bungeecord/src/main/java/com/vexsoftware/votifier/bungee/NuVotifier.java
@@ -306,7 +306,7 @@ public class NuVotifier extends Plugin implements VoteHandler, VotifierPlugin {
 
                 if (token != null) {
                     ProxyForwardingVoteSource.BackendServer server = new ProxyForwardingVoteSource.BackendServer(s,
-                            new InetSocketAddress(address, section.getShort("port")),
+                            new InetSocketAddress(address, section.getInt("port")),
                             token);
                     serverList.add(server);
                 }
